Implement batching
and try out a more readable target decomposition convention.
prop_GPT now supports taking in multiple training samples along the batch
dimension, and the handling of different samples is handled more explicitly.
This will make it easier to do circuit analysis on multiple examples in
a dataset at once, and also enables us to read the output more easily.

Also update the GPT circuit discovery notebook. The most important thing
is that we don't pad the sequence length up to 512. For the IOI dataset
things are more like 16-20 tokens long, and the difference in runtime is
like 20x.

Also note I measured: running the decompositions in batches of 64 is
approximately 4 times faster (on my laptop).

Also fix a bug: previously was taking mean over the ioi_dataset, not the
abc_dataset