Implement alternative algorithm; more sanity checks
Currently running into problems getting the standard algorithm to work
(the most relevant nodes to the target set end up being ones very early
in the network, leading to a not very interesting circuit and definitely
not the one found in the IOI paper).

Try a new algorithm where we just find the most relevant nodes in every layer
and then weight the target scores according to the relevance scores that the
target nodes had in the previous iteration, then choose the nodes that had the
highest relevances overall during the runtime of the algorithm.
The problem with this algorithm is that after not many iterations, the nodes
with the highest relevance scores are all at sequence position 0.