Complete analysis for docstring task
Maybe some graphs can still be made
Lot of work needs to be done on the patching functions to make
the "batch circuit analysis" work completely correctly.
That is, at every step of the algorithm we take the average
relevances across multiple datapoints, so that the result is stable.

Some hacks are in the HookedTransformer patching functions.
Since we don't want to patch with mean, but rather corrupted inputs,
we have to patch with a different entry per batch index.
This is very finicky, you have to remember to remove the hook every time
you're not running the model on the exact same batch of data.

Also complete the analysis in the docstring_exploration notebook.
To reproduce the result, run exactly the cells that are not collapsed in
the Setup section and then in the Experiments/All in one cell section
and then in the Experiments/Circuit Evaluation section.